Main Code1
This code schedules task according to the importance and d-day as well as considering the dependence.
This is a code part without topological sorting applied, but to have less time complexity, I've changed the dependence from a list input to int input to consider only 1 task before it should be allocated. (#heuristics)
After some fixing and trying diverse trials and versions, I found this is the most effective.(only considering non-topological sorting)
The heapsort is O(nlogn) time complexity, however, by having to cluster the activities considering dependencies and to generate a new value for such cluster, the O(n^2) time complexity was additionally taken by method 'cluster'. Thus, this whole code, when it runs for the first place it gets to have O(n^2) time complexity. However, this code is different from the ones from the start of the semester. Because it allows adding a new activity inside the existing list of tasks. This only takes O(n) time complexity in expense of memory O(n) for keeping the ready-made list after carefully considering the way to shorten the complexity from using nested for loop(or nested while loop) to a single for loop with if statements. As this is a code before getting to know the topological sorting, I've used already existing knowledge of simple coding and from the class(using queue structure) to achieve the goal of getting a result without an error. (#heuristics: availability)
Especially, considering the cluster was a technical challenge that I've had to face to build a greedy-optimal scheduler to have dependent activities be aligned with each other closely as well as having to sort them by this cluster.  (#gap analysis: knowing what I want to build and trying to find a break through from existing code from the start of the semester.)

(#cognitive persuasion by #evidencebased)
Time complexity has been calculated by :
 for loops/while loops as (n) or some what similar to this, it gets multiplied by other time complexities inside the loop. So, nested for loop with iterating n times each will be O(n^2). Without break, it will be big-Theta(n^2) in this example. 
if statement  and other single calculations are considered as constant.
These analysis was kept in mind as the versions upgraded to achieve the efficiency as well as achieving the wanted goal. So the newly added feature method 'new_list' was considered in expense of space(memory) the time complexity has decreased from O(n^2) to O(n). As a daily tasks are usually limited in daily life for individual human, it is unlikely that people reach the number of n that will take millions of years to run, however, we don't want the code to run more than a minute especially during the day when we are in middle of activity and just appending another task to do. We need this result to be quick to see what to do next. So, it is preferred to have smaller time complexity. Also as the computer is developing in hardware, the memory size has increased tremendously over time. Thus, having to make some memory use is ok for future use to decrease time complexity of the codes.